requirements.txt
# 后端依赖
Django>=4.2.0
djangorestframework>=3.14.0
mysqlclient>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.10.0
keras>=2.10.0
matplotlib>=3.7.0

# 前端依赖
# 前端依赖将在package.json中定义